Further code clean-up with added position and element removal functions.
Also made it so that the internal dynamic array is allocated as a block of memory without running any constructors.
This is more optimized, but it might be better for some types if their default constructor is run immediately upon allocation. Will consider undo-ing this change if it's more appropiate instead.