Implement separate-chaining hash table.
The implementation may need some improvements.
The copy and move semantics of linked lists was a big hurdle in implementing this. A lot of issues that had to be resolved along the way.
Feel free to look over the details of the implementation to consider some of the decisions made.