Implement resizing and reordering optimization.
It was very inefficient to actually grow the internal array (which involves allocating a new array, copying over elements, then deallocating the original array) only to replace it when reordering after constructing a new table.
This removes that inefficiency by manually computing the new needed capacity (with increaseCapacity()), then constructing the new table directly (we only grow the array directly if we don't have any entries to reorder in the first place).